Feature: Delete message

+++ src/Web/Features/Chat/Messages/DeleteMessage.cs
@@ -18,23 +18,51 @@ public sealed record DeleteMessage(Guid MessageId) : IRequest<Result>
     public sealed class Handler : IRequestHandler<DeleteMessage, Result>
     {
         private readonly IMessageRepository messageRepository;
+        private readonly IUnitOfWork unitOfWork;
+        private readonly ICurrentUserService currentUserService;
 
-        public Handler(IMessageRepository messageRepository)
+        public Handler(
+            IMessageRepository messageRepository, IUnitOfWork unitOfWork, ICurrentUserService currentUserService)
         {
             this.messageRepository = messageRepository;
+            this.unitOfWork = unitOfWork;
+            this.currentUserService = currentUserService;
         }
 
         public async Task<Result> Handle(DeleteMessage request, CancellationToken cancellationToken)
         {
             var message = await messageRepository.FindByIdAsync(request.MessageId, cancellationToken);
 
-             var deleted = await messageRepository
+            if(message is null)
+            {
+                return Result.Failure(Errors.Messages.MessageNotFound);
+            }
+
+            var userId = currentUserService.UserId;
+            var isAdmin = currentUserService.IsInRole("admin");
+
+            if(!isAdmin && message.CreatedById != userId) 
+            {
+                return Result.Failure(Errors.Messages.NotAllowedToDelete);
+            }
+
+            message.AddDomainEvent(new MessageDeleted(message.ChannelId, message.Id));
+
+            messageRepository.Remove(message);
+
+            await unitOfWork.SaveChangesAsync(cancellationToken);
+
+            return Result.Success();
+
+            /*
+            var deleted = await messageRepository
                 .GetAll(new MessageWithId(request.MessageId))
-                .ExecuteDeleteAsync();
+                .ExecuteDeleteAsync();      
 
             return deleted > 0 
                 ? Result.Success() 
                 : Result.Failure(Errors.Messages.MessageNotFound);
+            */
         }
     }
 }
